simplify with a split
