+++ packages/services/hmi-server/src/main/java/software/uncharted/terarium/hmiserver/models/petrinet/Species.java
@@ -4,7 +4,6 @@ import com.fasterxml.jackson.annotation.JsonAlias;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import lombok.Data;
-import software.uncharted.terarium.hmiserver.models.Ontology;
 
 import java.io.Serializable;
 import java.util.List;
