+++ packages/services/hmi-server/src/main/java/software/uncharted/terarium/hmiserver/models/petrinet/SpeciesJsonDeserializer.java
@@ -5,13 +5,9 @@ import com.fasterxml.jackson.core.JsonParser;
 import com.fasterxml.jackson.databind.DeserializationContext;
 import com.fasterxml.jackson.databind.JsonDeserializer;
 import com.fasterxml.jackson.databind.JsonNode;
-import software.uncharted.terarium.hmiserver.models.Ontology;
 
-import javax.persistence.criteria.Root;
 import java.io.IOException;
 import java.util.ArrayList;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 
 public class SpeciesJsonDeserializer extends JsonDeserializer<Species> {
 	/*
@@ -22,33 +18,23 @@ public class SpeciesJsonDeserializer extends JsonDeserializer<Species> {
 		}
 	 */
 
-
 	@Override
 	public Species deserialize(final JsonParser p, final DeserializationContext ctxt) throws IOException, JacksonException {
 		JsonNode node = p.getCodec().readTree(p);
 		final Species species = new Species();
+
 		species.setName(node.get("sname").asText());
-
-		final Pattern pattern = Pattern.compile("\\(\\'\\w+\\', \\'\\w+:\\d+\\'\\)");
-
-		final Matcher miraIdsMatcher = pattern.matcher(node.get("mira_ids").asText());
 		species.setMiraIds(new ArrayList<>());
-		int i = miraIdsMatcher.groupCount();
-		if (i > 0) {
-			while (miraIdsMatcher.find()) {
-				species.getMiraIds().add(new Ontology(miraIdsMatcher.group(i)));
-				i++;
-			}
+		species.setMiraContext(new ArrayList<>());
+
+		final String[] mira_ids = node.get("mira_ids").asText().split("\\), \\(");
+		for (String ontology : mira_ids) {
+			species.getMiraIds().add(new Ontology(ontology));
 		}
 
-		final Matcher miraContextMatcher = pattern.matcher(node.get("mira_context").asText());
-		species.setMiraContext(new ArrayList<>());
-		i = miraContextMatcher.groupCount();
-		if (i > 0) {
-			while (miraContextMatcher.find()) {
-				species.getMiraContext().add(new Ontology(miraContextMatcher.group(i)));
-				i++;
-			}
+		final String[] mira_context = node.get("mira_context").asText().split("\\), \\(");
+		for (String ontology : mira_context) {
+			species.getMiraContext().add(new Ontology(ontology));
 		}
 
 		return species;
